\name{data}
\docType{data}
\alias{lansing}
\alias{polyb}
\title{Lansing Woods Trees}
\description{
  This is the Lansing Woods Tree data set and 
  an arbitrary polygon boundary within unit square. 
}
\usage{
  data(lansing)
  data(polyb)
}
\format{
  \code{lansing } is an \R \code{data.frame} with \code{x, y, marks}
  and \code{polyb} is an \R matrix with \code{x, y}.
}
\source{
  Lansing Woods trees from Lansing Woods, Clinton County, Michigan USA, 
  including hickories, maples, and oaks.
}
\note{
  The Lansing Woods tree data set and the arbitrary polygon are 
  provided here for the demonstration of the basic
  usage of this package. See Section \bold{Examples} in 
  \code{\link{spatialkernel-package}}.
}
\references{
  Gerrard, D.J. (1969) Competition quotient: a new measure of the competition 
  affecting individual forest trees. \emph{Research Bulletin 20}, 
  Agricultural Experiment Station, Michigan State University.
}
\keyword{datasets}
