% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metre.R
\name{metre}
\alias{metre}
\title{Plot Color Level Metre}
\usage{
metre(xl, yb, xr, yt, lab, cols = risk.colors(length(lab) - 1), shift = 0,
  cex = 1)
}
\arguments{
\item{xl}{Left coordinate of the color level metre bar.}

\item{yb}{Bottom coordinate of the color level metre bar.}

\item{xr}{Right coordinate of the color level metre bar.}

\item{yt}{Top coordinate of the color level metre bar.}

\item{lab}{metre level labels in the metre.}

\item{cols}{associated colours, defaults to use \code{\link{risk.colors}}.}

\item{shift}{distance to shift the label texts away from the metre bar.}

\item{cex}{numeric character expansion factor.}
}
\description{
This is a simple function provided here for the convenience of users. It adds
a key showing how the colors map to the values of an \code{image} plot.
}
\seealso{
\code{\link{risk.colors}}
}
\keyword{graphs}
\keyword{hplot}
