% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.version.R
\name{package.version}
\alias{package.version}
\title{Listing Loaded/Installed Package Versions
List version numbers of loaded or installed packages.}
\usage{
package.version(all.available = FALSE, lib.loc = NULL)
}
\arguments{
\item{all.available}{logical, if \code{TRUE} return all available packages.}

\item{lib.loc}{character vector describing the location of \R library trees
to search through, or \code{NULL}. The default value of \code{NULL} 
corresponds to all libraries currently known.}
}
\value{
A list with components
\describe{
  \item{package}{names of loaded or available packages 
    if \code{all.available} is \code{TRUE}.}
  \item{version}{associated package versions.}
}
}
\description{
Listing Loaded/Installed Package Versions
List version numbers of loaded or installed packages.
}
\details{
This is a revised version of the \pkg{base} \R function
  \code{.package}. It gives both package names and their version numbers.
}
\seealso{
\code{\link{.packages}}
}
\keyword{utilities}
