% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.colors.R
\name{risk.colors}
\alias{risk.colors}
\title{Color Palette}
\usage{
risk.colors(n)
}
\arguments{
\item{n}{number of colors (>= 1) to be in the palette.}
}
\description{
This color palette is designed to show risk levels with different colours 
  from small risk (bright orange) to high risk (dark red).
}
\examples{
  ## risk pie with ten levels
  pie(rep(1,10), labels = seq(0.1, 1, 0.1), col = risk.colors(10))
}
\seealso{
\code{\link{metre}}, \code{\link{colors}}, and \code{\link{palette}}.
}
\keyword{color}
\keyword{graphs}
